Animations Added
Added spinning and floating animations to collectibles some changes need to be made to prefab however as I can't (or at least can't figure out) how to get just the collectible model to move without the lighting or particle effects. Ideally the model would be an object childed to the collectible.