Animations Done + Also some bonus collectible fixes
Finished Animations touched with base with Lauren and Caleb. Updated the TestCollectible Prefab to add a dedicated child to the model so I can animate just the collectible model without making the lights and particles move in an unwanted manner. Finally made soem changes to the WearableCollectible script to address issues with the hats displaying in the level.