Canvas Gui Dialog Box Art & Tweaks
Added art onto the canvas gui dialog box

shifted it into the bottom left corner to fit the UI vision. Should be shifted up enough to not interfere with checkpoint markers.

Had to mess with some of the anti-aliasing on the sprite sheet, so thats why it has funny compression settings. otherwise it gets phantom lines between the pieces.